Add trigger message and remove remote stop
