+++ 03_Modules/Configuration/src/index.ts
@@ -3,6 +3,8 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { ConfigurationModuleApi } from './module/api';
+export { ConfigurationModuleApi as ConfigurationDataApi } from './module/DataApi';
+export { ConfigurationModuleApi as ConfigurationOcpp201Api } from './module/Ocpp201Api';
+export { ConfigurationModuleApi as ConfigurationOcpp16Api } from './module/Ocpp16Api';
 export { IConfigurationModuleApi } from './module/interface';
 export { ConfigurationModule } from './module/module';
