+++ 03_Modules/EVDriver/src/module/api.ts
@@ -3,14 +3,17 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-import { FastifyInstance } from 'fastify';
+import { FastifyInstance, FastifyRequest } from 'fastify';
 import { ILogObj, Logger } from 'tslog';
 import { IEVDriverModuleApi } from './interface';
 import { EVDriverModule } from './module';
 import {
   AbstractModuleApi,
+  AsDataEndpoint,
   AsMessageEndpoint,
+  AuthorizationDataSchema,
   CallAction,
+  HttpMethod,
   IMessageConfirmation,
   OCPP2_0_1_Namespace,
   OCPP1_6_Namespace,
@@ -20,7 +23,14 @@ import {
   OCPPVersion,
 } from '@citrineos/base';
 import {
+  AuthorizationQuerySchema,
+  AuthorizationQuerystring,
+  AuthorizationRestrictions,
+  AuthorizationRestrictionsSchema,
   CallMessage,
+  ChargingStationKeyQuerySchema,
+  ChargingStationKeyQuerystring,
+  LocalListVersion,
 } from '@citrineos/data';
 import { validateChargingProfileType } from '@citrineos/util';
 import { v4 as uuidv4 } from 'uuid';
@@ -43,6 +53,92 @@ export class EVDriverModuleApi
   ) {
     super(evDriverModule, server, logger);
   }
+
+  /**
+   * Data Endpoint Methods
+   */
+
+  @AsDataEndpoint(
+    OCPP2_0_1_Namespace.AuthorizationData,
+    HttpMethod.Put,
+    AuthorizationQuerySchema,
+    AuthorizationDataSchema,
+  )
+  putAuthorization(
+    request: FastifyRequest<{
+      Body: OCPP2_0_1.AuthorizationData;
+      Querystring: AuthorizationQuerystring;
+    }>,
+  ): Promise<OCPP2_0_1.AuthorizationData | undefined> {
+    return this._module.authorizeRepository.createOrUpdateByQuerystring(
+      request.body,
+      request.query,
+    );
+  }
+
+  @AsDataEndpoint(
+    OCPP2_0_1_Namespace.AuthorizationRestrictions,
+    HttpMethod.Put,
+    AuthorizationQuerySchema,
+    AuthorizationRestrictionsSchema,
+  )
+  putAuthorizationRestrictions(
+    request: FastifyRequest<{
+      Body: AuthorizationRestrictions;
+      Querystring: AuthorizationQuerystring;
+    }>,
+  ): Promise<OCPP2_0_1.AuthorizationData[]> {
+    return this._module.authorizeRepository.updateRestrictionsByQuerystring(
+      request.body,
+      request.query,
+    );
+  }
+
+  @AsDataEndpoint(
+    OCPP2_0_1_Namespace.AuthorizationData,
+    HttpMethod.Get,
+    AuthorizationQuerySchema,
+  )
+  getAuthorization(
+    request: FastifyRequest<{ Querystring: AuthorizationQuerystring }>,
+  ): Promise<OCPP2_0_1.AuthorizationData[]> {
+    return this._module.authorizeRepository.readAllByQuerystring(request.query);
+  }
+
+  @AsDataEndpoint(
+    OCPP2_0_1_Namespace.AuthorizationData,
+    HttpMethod.Delete,
+    AuthorizationQuerySchema,
+  )
+  deleteAuthorization(
+    request: FastifyRequest<{ Querystring: AuthorizationQuerystring }>,
+  ): Promise<string> {
+    return this._module.authorizeRepository
+      .deleteAllByQuerystring(request.query)
+      .then(
+        (deletedCount) =>
+          deletedCount.toString() +
+          ' rows successfully deleted from ' +
+          OCPP2_0_1_Namespace.AuthorizationData,
+      );
+  }
+
+  @AsDataEndpoint(
+    OCPP2_0_1_Namespace.LocalListVersion,
+    HttpMethod.Get,
+    ChargingStationKeyQuerySchema,
+  )
+  async getLocalAuthorizationListVersion(
+    request: FastifyRequest<{ Querystring: ChargingStationKeyQuerystring }>,
+  ): Promise<LocalListVersion | undefined> {
+    return await this._module.localAuthListRepository.readOnlyOneByQuery({
+      stationId: request.query.stationId,
+    });
+  }
+
+  /**
+   * Message Endpoint Methods
+   */
 
   @AsMessageEndpoint(
     OCPP2_0_1_CallAction.RequestStartTransaction,
@@ -446,5 +542,19 @@ export class EVDriverModuleApi
   protected _toMessagePath(input: CallAction): string {
     const endpointPrefix = this._module.config.modules.evdriver.endpointPrefix;
     return super._toMessagePath(input, endpointPrefix);
+  }
+
+  /**
+   * Overrides superclass method to generate the URL path based on the input {@link Namespace}
+   * and the module's endpoint prefix configuration.
+   *
+   * @param {Namespace} input - The input {@link Namespace}.
+   * @return {string} - The generated URL path.
+   */
+  protected _toDataPath(
+    input: OCPP2_0_1_Namespace | OCPP1_6_Namespace | Namespace,
+  ): string {
+    const endpointPrefix = this._module.config.modules.evdriver.endpointPrefix;
+    return super._toDataPath(input, endpointPrefix);
   }
 }
